Reorganize into pysac subpackage.
