+++ obspy/io/sac/pysac/__init__.py
@@ -1,0 +1,1 @@
+from .sactrace import SACTrace
