+++ obspy/io/sac/pysac/header.py
